ParseQri_Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pandas==2.2.3 
numpy==1.26.3
passlib==1.7.4
email-validator==2.2.0
python-multipart==0.0.20
redis==5.0.1








